Reverted to previous version
It felt like I was repeating the same thing in the Server and the Websocket, so I tried to eliminate repeat code. The first thing I did is try to get the Websocket to call the Server Facade, but this did not work because Server Facade requires you to start a server and the tests anticipate calling Websocket methods without having the Server set up. Then I tried going around the Server completely, but I was unable to get any of the error messages from the Websocket to the Client. I can only get error messages from the Server to the CLient because they are wrapped in HTTP. I wasn't really able to figure out how to do it without doing the same thing in Websocket and the Server. It is not horrible to look at because I can factor out the repeating code, but it still will probably have to interact with SQL twice as many times; the program will take twice as long, which stinks because it is opening and closing a connection with SQL many many times and I think that is slowing it down.